use accessible view service
